refactor(dylint): extract use_tree_to_strings to lint_utils
Eliminate code duplication by moving use_tree_to_strings function from
DE0301, DE0308, and DE0706 lints to shared lint_utils crate.

Changes:
- Add use_tree_to_strings() to lint_utils with full documentation
- Update DE0301/DE0308 to use shared function (removed ~30 lines each)
- Update DE0706 to use shared function (removed ~24 lines)
- Add LintContext import where needed for EarlyContext methods
- Remove unused imports (EarlyContext, Crate, UseTree, UseTreeKind)

All tests pass for all three lints.

Addresses CodeRabbit code duplication feedback.
